# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SAFvsOil API",
    "description": "Product scaffold for SAF vs Oil data, scenarios, and assumptions.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/sqlite/market-alerts": {
      "get": {
        "tags": [
          "sqlite-alerts",
          "market-alerts-sqlite"
        ],
        "summary": "List Market Alerts",
        "description": "List market alerts with optional filtering.",
        "operationId": "list_market_alerts_v1_sqlite_market_alerts_get",
        "parameters": [
          {
            "name": "market_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by market type",
              "title": "Market Type"
            },
            "description": "Filter by market type"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status (active/inactive)",
              "title": "Status"
            },
            "description": "Filter by status (active/inactive)"
          },
          {
            "name": "db_path",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "/opt/safvsoil/data/market.db",
              "title": "Db Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MarketAlertRead"
                  },
                  "title": "Response List Market Alerts V1 Sqlite Market Alerts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "sqlite-alerts",
          "market-alerts-sqlite"
        ],
        "summary": "Create Market Alert",
        "description": "Create new market alert.",
        "operationId": "create_market_alert_v1_sqlite_market_alerts_post",
        "parameters": [
          {
            "name": "db_path",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "/opt/safvsoil/data/market.db",
              "title": "Db Path"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketAlertCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketAlertRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sqlite/market-alerts/{alert_id}": {
      "get": {
        "tags": [
          "sqlite-alerts",
          "market-alerts-sqlite"
        ],
        "summary": "Get Market Alert",
        "description": "Get specific market alert by ID.",
        "operationId": "get_market_alert_v1_sqlite_market_alerts__alert_id__get",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alert Id"
            }
          },
          {
            "name": "db_path",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "/opt/safvsoil/data/market.db",
              "title": "Db Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketAlertRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "sqlite-alerts",
          "market-alerts-sqlite"
        ],
        "summary": "Update Market Alert",
        "description": "Update market alert.",
        "operationId": "update_market_alert_v1_sqlite_market_alerts__alert_id__put",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alert Id"
            }
          },
          {
            "name": "db_path",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "/opt/safvsoil/data/market.db",
              "title": "Db Path"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketAlertUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketAlertRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sqlite-alerts",
          "market-alerts-sqlite"
        ],
        "summary": "Delete Market Alert",
        "description": "Delete market alert.",
        "operationId": "delete_market_alert_v1_sqlite_market_alerts__alert_id__delete",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alert Id"
            }
          },
          {
            "name": "db_path",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "/opt/safvsoil/data/market.db",
              "title": "Db Path"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sqlite/market-alerts/{alert_id}/trigger": {
      "put": {
        "tags": [
          "sqlite-alerts",
          "market-alerts-sqlite"
        ],
        "summary": "Trigger Market Alert",
        "description": "Mark alert as triggered with current timestamp.",
        "operationId": "trigger_market_alert_v1_sqlite_market_alerts__alert_id__trigger_put",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alert Id"
            }
          },
          {
            "name": "db_path",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "/opt/safvsoil/data/market.db",
              "title": "Db Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketAlertRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MarketAlertCreate": {
        "properties": {
          "market_type": {
            "type": "string",
            "title": "Market Type",
            "description": "Market type: ARA, US_Gulf, EU_ETS"
          },
          "threshold_type": {
            "type": "string",
            "title": "Threshold Type",
            "description": "Type: 'above' or 'below'"
          },
          "threshold_value": {
            "type": "number",
            "title": "Threshold Value"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Status: active or inactive",
            "default": "active"
          }
        },
        "type": "object",
        "required": [
          "market_type",
          "threshold_type",
          "threshold_value"
        ],
        "title": "MarketAlertCreate",
        "description": "Schema for creating market alert."
      },
      "MarketAlertRead": {
        "properties": {
          "market_type": {
            "type": "string",
            "title": "Market Type",
            "description": "Market type: ARA, US_Gulf, EU_ETS"
          },
          "threshold_type": {
            "type": "string",
            "title": "Threshold Type",
            "description": "Type: 'above' or 'below'"
          },
          "threshold_value": {
            "type": "number",
            "title": "Threshold Value"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Status: active or inactive",
            "default": "active"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "last_triggered": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Triggered"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "market_type",
          "threshold_type",
          "threshold_value",
          "id",
          "last_triggered",
          "created_at",
          "updated_at"
        ],
        "title": "MarketAlertRead",
        "description": "Schema for reading market alert."
      },
      "MarketAlertUpdate": {
        "properties": {
          "threshold_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold Type"
          },
          "threshold_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold Value"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "title": "MarketAlertUpdate",
        "description": "Schema for updating market alert."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}